requirements.txt
# Core ML and AI libraries
torch>=2.0.0
transformers>=4.30.0
whisper-openai>=20230314
librosa>=0.10.0
scipy>=1.10.0
numpy>=1.24.0

# Computer Vision
opencv-python>=4.8.0
mediapipe>=0.10.0
face-recognition>=1.3.0

# Speech Processing
pyaudio>=0.2.11
pydub>=0.25.1
webrtcvad>=2.0.10
python-speech-features>=0.6

# NLP and Sentiment Analysis
spacy>=3.6.0
textblob>=0.17.1
vaderSentiment>=3.3.2
nltk>=3.8.1

# Web Framework and API
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
aiofiles>=23.1.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
jinja2>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0
